pid parser for 215, proper pid priority setting

+++ app/src/main/java/com/pitstop/models/TripIndicator.java
@@ -10,10 +10,10 @@ import com.pitstop.network.RequestCallback;
 
 public abstract class TripIndicator {
 
-    int tripId;
+    long tripId;
     String rtcTime;
 
-    TripIndicator(int tripId, String rtcTime) {
+    TripIndicator(long tripId, String rtcTime) {
         this.tripId = tripId;
         this.rtcTime = rtcTime;
     }
@@ -22,7 +22,7 @@ public abstract class TripIndicator {
         return rtcTime;
     }
 
-    public int getTripId() {
+    public long getTripId() {
         return tripId;
     }
