+++ app/src/main/java/com/pitstop/utils/NetworkHelper.java
@@ -354,7 +354,7 @@ public class NetworkHelper {
         postNoAuth("scan/trip", callback, body);
     }
 
-    public void saveTripMileage(int tripId, String mileage, String rtcTime, RequestCallback callback) {
+    public void saveTripMileage(long tripId, String mileage, String rtcTime, RequestCallback callback) {
         LOGI(TAG, String.format("saveTripMileage: tripId: %s," +
                 " mileage: %s, rtcTime: %s", tripId, mileage, rtcTime));
